Break out test consumer callback for message testing
